requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
openai>=1.55.3
streamlit==1.28.1
requests==2.31.0
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
plotly==5.17.0
streamlit-drawable-canvas==0.9.3
httpx==0.27.2 